refactor: redesign from scratch
