+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "php-tmdb/laravel",
-    "description": "Laravel Package for TMDB ( The Movie Database ) API. Provides easy access to the wtfzdotnet/php-tmdb-api library.",
+    "description": "Laravel Package for TMDB ( The Movie Database ) API. Provides easy access to the php-tmdb/api library.",
     "keywords": ["Laravel", "tmdb", "api", "php", "wrapper", "movie", "tv", "tv show", "tvdb"],
     "authors": [
         {
@@ -10,16 +10,17 @@
     ],
     "license": "MIT",
     "require": {
-        "php": ">=5.4.0",
-        "illuminate/support": "~5.0 || ~4.0",
-        "php-tmdb/api": "~2.0"
+        "php": ">=7.1.3",
+        "php-tmdb/api": "~3.0",
+        "illuminate/support": "^6|^7|^8",
+        "symfony/event-dispatcher": ">=4.4,<6"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.5",
-        "symfony/event-dispatcher": "~2.8|~3.0",
-        "illuminate/events": "~5.0",
+        "phpunit/phpunit": "^8.0|^9.0",
+        "illuminate/events": "^6|^7|^8",
         "doctrine/cache": ">=1.3.0",
-        "monolog/monolog": ">=1.7.0"
+        "monolog/monolog": "^1.12|^2.0",
+        "orchestra/testbench": "~3.6.7 || ~3.7.8 || ~3.8.6 || ^4.8 || ^5.2 || ^6.0"
     },
     "suggest": {
         "doctrine/cache": "Required if you want to make use of caching features.",
@@ -28,6 +29,11 @@
     "autoload": {
         "psr-4": {
             "Tmdb\\Laravel\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tmdb\\Laravel\\Tests\\": "tests/"
         }
     },
     "minimum-stability": "stable",
